Refactor AtmosphereLUTSystem into multiple focused source files
Split the ~1573-line AtmosphereLUTSystem.cpp into 6 files organized by responsibility:

- AtmosphereLUTSystem.cpp: Core init/destroy orchestration (~105 lines)
- AtmosphereLUTResources.cpp: LUT image/buffer/sampler creation (~310 lines)
- AtmosphereLUTDescriptors.cpp: Descriptor layouts and sets (~480 lines)
- AtmosphereLUTPipelines.cpp: Compute pipeline creation (~180 lines)
- AtmosphereLUTCompute.cpp: Compute dispatch methods (~200 lines)
- AtmosphereLUTExport.cpp: PNG export functionality (~200 lines)

This improves code maintainability by separating concerns into smaller,
focused units while preserving all existing functionality.
